+++ CHANGES.md
@@ -2,3 +2,12 @@
 * Codebase has been changed to es6.
 * Use of import and javascript latest features
 * Now, bundle file can be used using script tag
+
+
+# Added Currency Symbol Maps
+* removed External Dependency for currency-symbol-map. Now we have 4 methods
+*  currencySymbolMap => returns list of currency with name, number, symbol
+*  getSymbolFromCurrency,
+*  getNameFromCurrency,
+*  getSafeSymbolFromCurrency,
+*  getSafeNameFromCurrency